refactor : 카테고리 도메인
1. reformat
2. optimize imports
