+++ src/main/java/com/sparta/bookflex/domain/category/controller/CategoryController.java
@@ -1,18 +1,24 @@
-package com.sparta.bookflex.domain.category;
+package com.sparta.bookflex.domain.category.controller;
 
 import com.sparta.bookflex.common.aop.Envelop;
+import com.sparta.bookflex.domain.category.dto.CategoryAllResponseDto;
+import com.sparta.bookflex.domain.category.dto.CategoryNameResponseDto;
+import com.sparta.bookflex.domain.category.service.CategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/categories")
-public class CategoryEnumController {
-    private final CategoryEnumService categoryEnumService;
+public class CategoryController {
+    private final CategoryService categoryService;
 
     /*모든 카테고리 목록 조회
     GET /categories
@@ -29,9 +35,9 @@ public class CategoryEnumController {
     * */
     @GetMapping("/enum")
     @Envelop("모든 카테고리 조회에 성공하였습니다.")
-    public ResponseEntity<List<CategoryEnumResponseDto>> getAllCategories() {
+    public ResponseEntity<List<CategoryAllResponseDto>> getAllCategories() {
 
-        List<CategoryEnumResponseDto> responseList = categoryEnumService.getAllCategories();
+        List<CategoryAllResponseDto> responseList = categoryService.getAllCategories();
         return ResponseEntity.status(HttpStatus.OK).body(responseList);
     }
 
@@ -39,7 +45,7 @@ public class CategoryEnumController {
     @Envelop("메인 카테고리 조회에 성공하였습니다.")
     public ResponseEntity<List<CategoryNameResponseDto>> getAllMainCategories() {
 
-        List<CategoryNameResponseDto> responseList = categoryEnumService.getAllMainCategories();
+        List<CategoryNameResponseDto> responseList = categoryService.getAllMainCategories();
         return ResponseEntity.status(HttpStatus.OK).body(responseList);
     }
 
@@ -59,7 +65,7 @@ public class CategoryEnumController {
     @Envelop("서브 카테고리 목록 조회에 성공하였습니다.")
     public ResponseEntity<List<CategoryNameResponseDto>> getAllSubCategories(@RequestParam(value = "mainCategory", defaultValue = "카테고리") String mainCategory) {
 
-        List<CategoryNameResponseDto> responseList = categoryEnumService.getAllSubCategories(mainCategory);
+        List<CategoryNameResponseDto> responseList = categoryService.getAllSubCategories(mainCategory);
         return ResponseEntity.status(HttpStatus.OK).body(responseList);
     }
 }
